Add all the necessary code for the project
This has adds all of the code for the project. It sets up the
variables for user input as well as computer generated choice. It
sets up variable for computer and player scores.

When player inputs the answer, the program checks if it is a valid
answer and capitalized it. Then it stores it in a variable. All
of this is part of the separate function.

Computer generated choice is part of a separate function. It is done
based on the random number from 0 to 2. Each choice is assigned rock,
paper or scissors. The result is stored in the variable.

There is a function that checks who won by comparing a value in user
variable and computer variable. The function then returns a winner and
increments the score of whoever won.

The game function has a 5 round loop and on each loop is runs the game
and console logs the results for the round. At the end of the game
displays the game winner.